classify logical cte table into a separate file
