+++ pkg/planner/core/BUILD.bazel
@@ -24,6 +24,7 @@ go_library(
         "logical_aggregation.go",
         "logical_apply.go",
         "logical_cte.go",
+        "logical_cte_table.go",
         "logical_datasource.go",
         "logical_expand.go",
         "logical_index_scan.go",
