+++ pkg/planner/core/logical_plans.go
@@ -22,9 +22,7 @@ import (
 	"github.com/pingcap/tidb/pkg/parser/auth"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/planner/core/base"
-	"github.com/pingcap/tidb/pkg/planner/core/operator/logicalop"
 	fd "github.com/pingcap/tidb/pkg/planner/funcdep"
-	"github.com/pingcap/tidb/pkg/planner/property"
 	"github.com/pingcap/tidb/pkg/planner/util"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/types"
@@ -292,66 +290,4 @@ func (s *ShowContents) MemoryUsage() (sum int64) {
 	sum = emptyShowContentsSize + int64(len(s.DBName)) + s.Partition.MemoryUsage() + s.IndexName.MemoryUsage() +
 		int64(cap(s.Roles))*size.SizeOfPointer
 	return
-}
-
-// CTEClass holds the information and plan for a CTE. Most of the fields in this struct are the same as cteInfo.
-// But the cteInfo is used when building the plan, and CTEClass is used also for building the executor.
-type CTEClass struct {
-	// The union between seed part and recursive part is DISTINCT or DISTINCT ALL.
-	IsDistinct bool
-	// seedPartLogicalPlan and recursivePartLogicalPlan are the logical plans for the seed part and recursive part of this CTE.
-	seedPartLogicalPlan      base.LogicalPlan
-	recursivePartLogicalPlan base.LogicalPlan
-	// seedPartPhysicalPlan and recursivePartPhysicalPlan are the physical plans for the seed part and recursive part of this CTE.
-	seedPartPhysicalPlan      base.PhysicalPlan
-	recursivePartPhysicalPlan base.PhysicalPlan
-	// storageID for this CTE.
-	IDForStorage int
-	// optFlag is the optFlag for the whole CTE.
-	optFlag   uint64
-	HasLimit  bool
-	LimitBeg  uint64
-	LimitEnd  uint64
-	IsInApply bool
-	// pushDownPredicates may be push-downed by different references.
-	pushDownPredicates []expression.Expression
-	ColumnMap          map[string]*expression.Column
-	isOuterMostCTE     bool
-}
-
-const emptyCTEClassSize = int64(unsafe.Sizeof(CTEClass{}))
-
-// MemoryUsage return the memory usage of CTEClass
-func (cc *CTEClass) MemoryUsage() (sum int64) {
-	if cc == nil {
-		return
-	}
-
-	sum = emptyCTEClassSize
-	if cc.seedPartPhysicalPlan != nil {
-		sum += cc.seedPartPhysicalPlan.MemoryUsage()
-	}
-	if cc.recursivePartPhysicalPlan != nil {
-		sum += cc.recursivePartPhysicalPlan.MemoryUsage()
-	}
-
-	for _, expr := range cc.pushDownPredicates {
-		sum += expr.MemoryUsage()
-	}
-	for key, val := range cc.ColumnMap {
-		sum += size.SizeOfString + int64(len(key)) + size.SizeOfPointer + val.MemoryUsage()
-	}
-	return
-}
-
-// LogicalCTETable is for CTE table
-type LogicalCTETable struct {
-	logicalop.LogicalSchemaProducer
-
-	seedStat     *property.StatsInfo
-	name         string
-	idForStorage int
-
-	// seedSchema is only used in columnStatsUsageCollector to get column mapping
-	seedSchema *expression.Schema
 }
